CS/Typos

+++ inc/deploygroup_staticdata.class.php
@@ -121,6 +121,7 @@ class PluginGlpiinventoryDeployGroup_Staticdata extends CommonDBRelation
     */
     public static function displayTabContentForItem(CommonGLPI $item, $tabnum = 1, $withtemplate = 0)
     {
+        /** @var PluginGlpiinventoryDeployGroup $item */
         switch ($tabnum) {
             case 1:
                 self::showCriteriaAndSearch($item);
@@ -372,7 +373,7 @@ class PluginGlpiinventoryDeployGroup_Staticdata extends CommonDBRelation
         if (isset($files_data['importcsvfile']['tmp_name'])) {
             if (($handle = fopen($files_data['importcsvfile']['tmp_name'], "r")) !== false) {
                 while (($data = fgetcsv($handle, 1000, $_SESSION["glpicsv_delimiter"], '"', '')) !== false) {
-                    $input['items_id'] = str_replace(' ', '', $data[0]);
+                    $input['items_id'] = (int)str_replace(' ', '', $data[0]);
                     if ($computer->getFromDB($input['items_id'])) {
                         $pfDeployGroup_static->add($input);
                     }
